🔄️ Rename legacy lifestream package and wire scheduler to new importers
- Rename imports/lifestream/ → imports/lifestream_legacy/ to prevent
  sys.path collision with the installed src/lifestream/ package
- Update all legacy import scripts to use lifestream_legacy
- Wire scheduler.py to use src/lifestream/core/ directly instead of
  the legacy jobs.py (which lacked new-style IMPORTERS dispatch)
- Remove legacy scripts that have been converted to new-style importers:
  atom, flickr, github_commits, lastfm, mastodon_toots, steam,
  switchbot, wordpress
- Fix F402: rename loop variable config → job_config in list_jobs()
