+++ contrib/oyster_import_csv.py
@@ -7,8 +7,7 @@ import hashlib
 from datetime import datetime
 
 import pytz
-
-from lifestream.db import get_connection, get_cursor
+from lifestream_legacy.db import get_connection, get_cursor
 
 dbcxn = get_connection()
 cursor = get_cursor(dbcxn)
