--- imports/atom.py
@@ -1,49 +1,0 @@
-"""Generic Atom/RSS feed importer for Lifestream."""
-
-# Python
-import logging
-from datetime import datetime
-from time import mktime
-
-# Libraries
-import feedparser
-
-# Local
-import lifestream
-from lifestream.db import EntryStore
-
-logger = logging.getLogger("Atom")
-
-# Add custom arguments before calling parse_args
-lifestream.arguments.add_argument("type")
-lifestream.arguments.add_argument("url")
-
-
-def main():
-    """Import entries from an Atom/RSS feed."""
-    args = lifestream.parse_args()
-
-    entry_store = EntryStore()
-
-    logger.info("Grabbing %s", args.url)
-    fp = feedparser.parse(args.url)
-
-    for i in range(len(fp["entries"])):
-        o_item = fp["entries"][i]
-        item_id = o_item["guid"]
-        dt = datetime.fromtimestamp(mktime(o_item["updated_parsed"]))
-        updated = dt.strftime("%Y-%m-%d %H:%M")
-        logger.info("Adding new %s item: %s", args.type, o_item["title"])
-
-        entry_store.add_entry(
-            type=args.type,
-            id=item_id,
-            title=o_item["title"],
-            source=args.type,
-            date=updated,
-            url=o_item["links"][0]["href"],
-        )
-
-
-if __name__ == "__main__":
-    main()
