+++ imports/destiny2.py
@@ -9,17 +9,17 @@ import time
 from datetime import datetime, timedelta
 
 import dateutil.parser
+
+# Local
+import lifestream_legacy as lifestream
 import pytz
 
 # Libraries
 import requests
-
-# Local
-import lifestream
-from lifestream import code_fetcher as CodeFetcher9000
-from lifestream import destiny_exceptions
-from lifestream.cache import check_and_set_backoff, set_backoff, should_backoff
-from lifestream.db import EntryStore
+from lifestream_legacy import code_fetcher as CodeFetcher9000
+from lifestream_legacy import destiny_exceptions
+from lifestream_legacy.cache import check_and_set_backoff, set_backoff, should_backoff
+from lifestream_legacy.db import EntryStore
 
 entry_store = EntryStore()
