+++ imports/gw2.py
@@ -3,16 +3,15 @@ import logging
 import sys
 from datetime import datetime
 
+# Local
+import lifestream_legacy as lifestream
 import requests
 from guildwars2api.base import GuildWars2APIError
 
 # Libraries
 from guildwars2api.v2 import GuildWars2API
-
-# Local
-import lifestream
-from lifestream.cache import check_and_set_backoff, file_cache
-from lifestream.db import EntryStore
+from lifestream_legacy.cache import check_and_set_backoff, file_cache
+from lifestream_legacy.db import EntryStore
 
 entry_store = EntryStore()
