+++ imports/oyster.py
@@ -8,14 +8,13 @@ import sys
 from datetime import datetime
 from time import sleep
 
+# Local
+import lifestream_legacy as lifestream
 import pytz
+from lifestream_legacy.db import EntryStore
 
 # Libraries
 from mechanize import Browser, RobustFactory
-
-# Local
-import lifestream
-from lifestream.db import EntryStore
 
 logger = logging.getLogger("Oyster")
 args = lifestream.parse_args()
