+++ imports/tweets.py
@@ -6,15 +6,15 @@ import sys
 from urllib.error import URLError
 
 import dateutil.parser
+
+# Local
+import lifestream_legacy as lifestream
 import pytz
 
 # Libraries
 import twitter
-
-# Local
-import lifestream
-from lifestream.db import EntryStore
-from lifestream.oauth_utils import read_token_file
+from lifestream_legacy.db import EntryStore
+from lifestream_legacy.oauth_utils import read_token_file
 
 entry_store = EntryStore()
