+++ scheduler.py
@@ -22,7 +22,6 @@ import signal
 import sys
 from datetime import datetime
 
-# Parse arguments BEFORE importing lifestream (which has its own argparse setup)
 parser = argparse.ArgumentParser(
     description="Lifestream Scheduler",
     formatter_class=argparse.RawDescriptionHelpFormatter,
@@ -32,24 +31,20 @@ parser.add_argument("--list", action="store_true", help="List configured jobs")
 parser.add_argument("--status", action="store_true", help="Show job status")
 parser.add_argument("--run", metavar="JOB", help="Run a specific job immediately")
 parser.add_argument("--debug", action="store_true", help="Enable debug logging")
-SCHEDULER_ARGS, remaining_args = parser.parse_known_args()
-
-# Replace sys.argv with remaining args so lifestream's argparse doesn't break
-sys.argv = [sys.argv[0]] + remaining_args
+SCHEDULER_ARGS, _ = parser.parse_known_args()
 
 from apscheduler.executors.pool import ThreadPoolExecutor  # noqa: E402
 from apscheduler.jobstores.redis import RedisJobStore  # noqa: E402
 from apscheduler.schedulers.blocking import BlockingScheduler  # noqa: E402
 from apscheduler.triggers.cron import CronTrigger  # noqa: E402
 
-# Set up path for imports - lifestream expects to run from imports/ directory
+# Add imports/ to path for legacy script fallback (run_import falls back to importlib)
 basedir = os.path.dirname(os.path.abspath(__file__))
 imports_dir = os.path.join(basedir, "imports")
-os.chdir(imports_dir)  # Change to imports/ so lifestream finds config correctly
-sys.path.insert(0, imports_dir)
+sys.path.append(imports_dir)
 
-import lifestream  # noqa: E402
-from lifestream.jobs import run_import, run_shell_command  # noqa: E402
+from lifestream.core.config import config  # noqa: E402
+from lifestream.core.jobs import run_import, run_shell_command  # noqa: E402
 
 logger = logging.getLogger("Scheduler")
 
@@ -92,11 +87,11 @@ def get_schedules():
     """Read schedules from config.ini [schedules] section."""
     schedules = {}
 
-    if not lifestream.config.has_section("schedules"):
+    if not config.has_section("schedules"):
         logger.warning("No [schedules] section found in config.ini")
         return schedules
 
-    for job_name, cron_expr in lifestream.config.items("schedules"):
+    for job_name, cron_expr in config.items("schedules"):
         if not cron_expr or cron_expr.startswith(";") or cron_expr.startswith("#"):
             continue
 
@@ -120,10 +115,10 @@ def get_schedules():
 def create_scheduler():
     """Create and configure the APScheduler instance."""
     # Get Redis connection settings from config
-    redis_host = lifestream.config.get("redis", "host", fallback="localhost")
-    redis_port = int(lifestream.config.get("redis", "port", fallback=6379))
-    redis_username = lifestream.config.get("redis", "username", fallback=None)
-    redis_password = lifestream.config.get("redis", "password", fallback=None)
+    redis_host = config.get("redis", "host", fallback="localhost")
+    redis_port = int(config.get("redis", "port", fallback=6379))
+    redis_username = config.get("redis", "username", fallback=None)
+    redis_password = config.get("redis", "password", fallback=None)
 
     jobstores = {
         "default": RedisJobStore(
@@ -213,9 +208,9 @@ def list_jobs():
     print(f"{'Job Name':<25} {'Schedule':<20} {'Grace(s)':<10} {'Coalesce'}")
     print("-" * 70)
 
-    for job_name, config in sorted(schedules.items()):
+    for job_name, job_config in sorted(schedules.items()):
         print(
-            f"{job_name:<25} {config['cron']:<20} {config['misfire_grace_time']:<10} {config['coalesce']}"
+            f"{job_name:<25} {job_config['cron']:<20} {job_config['misfire_grace_time']:<10} {job_config['coalesce']}"
         )
